YOLO-FastestV2-AnchorFree/requirements.txt
matplotlib==3.5.1
numpy==1.23.4
opencv_python==4.5.5.64
Pillow==8.4.0
pycocotools==2.0.4
PyYAML==6.0
scikit_image==0.19.1
thop==0.1.1.post2209072238
torch==1.11.0
torchsummary==1.5.1
torchvision==0.12.0
tqdm==4.61.2
wandb==0.12.12
protobuf==3.19.0